Troubles with Manifest
After lots of trials, I am having some problems with the
AndroidManifest.xml. For some reason when I want to start a <activity,
it won’t allow me to link the activity with MainActivity. It has
something to do with the location of the files. Been trying to figure
this out all break. Gotten nowhere.